Create publications module
